requirements.txt
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
streamlit>=1.30
pytest>=7.0
